requirements.txt
discord.py
python-dotenv
aiohttp
Pillow
feedparser